Move connection log to internal package
